requirements.txt
Flask
pynacl
python-dotenv
flask-sqlalchemy
pymysql
cryptography
python-telegram-bot
